% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmNMF.R
\name{zm_nmf}
\alias{zm_nmf}
\title{Zero-masked NMF with angular regularization, and (in the future) sparse matrix support}
\usage{
zm_nmf(
  A,
  k = NULL,
  max.iter = 1000,
  rel.tol = 0.001,
  n.threads = 0,
  verbose = TRUE,
  inner.max.iter = 100,
  inner.rel.tol = 1e-06,
  trace = 5,
  alpha = 0
)
}
\arguments{
\item{A}{A matrix to be factorized, in dense format (currently, sparse support in the future)}

\item{k}{Decomposition rank, integer.}

\item{max.iter}{Maximum number of alternating NNLS solutions for H and W, integer. Default 1000.}

\item{rel.tol}{Stop criterion, defined as the relative tolerance between two successive iterations: |e2-e1|/avg(e1,e2). Default 1e-3.}

\item{n.threads}{Number of threads/CPUs to use. Default to 0 (all cores).}

\item{verbose}{Boolean, gives error updates for each trace iterations when TRUE.}

\item{inner.max.iter}{Maximum number of iterations passed to each inner W or H matrix updating function. Default 1e-6.}

\item{inner.rel.tol}{Stop criterion for the inner sequential coordinate descent least squares loop, defined as relative tolerance passed to inner W or H during matrix updating: |e2-e1|/avg(e1,e2). Default 1e-6.}

\item{trace}{An integer indicating how frequently the MSE error should be calculated and checked for convergence. To check error every iteration, specify 1. To avoid checking error at all, specify trace > max.iter.  Default 5.}

\item{alpha}{Angular penalty, ideal when factorizing symmetric matrices and W is expected to be nearly equal to H}
}
\value{
A list of W and H matrices
}
\description{
Zero-masked NMF with angular regularization, and (in the future) sparse matrix support
}
